+++ ninja-core/src/main/java/ninja/template/directives/TemplateEngineFreemarkerAuthenticityFormDirective.java
@@ -13,17 +13,13 @@ import freemarker.template.TemplateException;
 import freemarker.template.TemplateModel;
 import freemarker.template.TemplateModelException;
 
-/**
- * 
- * @author svenkubiak
- *
- */
+
 @SuppressWarnings("rawtypes")
 public class TemplateEngineFreemarkerAuthenticityFormDirective implements TemplateDirectiveModel {
-    private String authenticityToken;
+    private Context context;
 
     public TemplateEngineFreemarkerAuthenticityFormDirective(Context context) {
-        this.authenticityToken = context.getSession().getAuthenticityToken();
+        this.context = context;
     }
     
     @Override
@@ -37,6 +33,6 @@ public class TemplateEngineFreemarkerAuthenticityFormDirective implements Templa
         }
        
         Writer out = env.getOut();
-        out.append("<input type=\"hidden\" value=\"" + this.authenticityToken + "\" name=\"" + NinjaConstant.AUTHENTICITY_TOKEN +  "\" />");
+        out.append("<input type=\"hidden\" value=\"" + context.getSession().getAuthenticityToken() + "\" name=\"" + NinjaConstant.AUTHENTICITY_TOKEN +  "\" />");
     }
 }