+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,5 +1,7 @@
 Version X.X.X
 =============
+
+ * 2015-02-19 AuthenticityToken support (SecureFilter, template enhancements) (svenkubiak)
  * 2015-02-12 #301 Fixed system specific line separator for String comparison (raptaml)
  * 2012-02-03 minor refactoring in JaxyRoutes init (lukaseichler)
  * 2015-02-03 Added method to add and unset a cookie from context (svenkubiak)